Cleanup: Move static util methods to util class.
